backend/requirements.txt
aiofiles==23.1.0
anyio==4.11.0
asyncpg==0.30.0
click==8.3.0
fastapi==0.95.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
pydantic==1.10.24
python-dotenv==1.1.1
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.22.0
python-docx
reportlab
faster-whisper==1.0.3
transformers==4.46.3
sentencepiece==0.2.0
torch==2.2.2
reportlab==4.2.5
python-docx==1.1.2
langdetect==1.0.9
